b=1110 "stat" path if filesystem does not support d_type
If d_type != DT_DIR, it would assuming that the current
path pointed to a non directory file. This assumption is
incorrect if the filesystem does not support the d_type
field. In that case d_type == DT_UNKNOWN and a call to
"stat" or "lstat" is needed to check if the path points
to a directory or not.