Add pacemaker configuration script, ldev2cib, and ocf resource agents
This is the first pass at these scripts.  They are not yet in finished
form, but basic functionality is in place.  Notably they are lacking
in comments in places.

The "ldev2cib" script takes an ldev.conf file and produces the XML for
a cib.xml file (pacemaker's configuration file) on stdout.

The "lustre" and "zpool" scripts are OCF Resource Agent scripts intended
to be used by pacemaker.  The afor mentioned ldev2cib script assumes
these resource agents are installed.