ldev2cib: make skip_pattern compatible with python-2.6.6 and 2.7.5
With python-2.6.6 and 2.7.5, skip_pattern fails to compile:

Traceback (most recent call last):
  File "./ldev2cib", line 107, in <module>
    skip_pattern = re.compile('^(\s*)?(#.*)?\n?$')
  File "/usr/lib64/python2.6/re.py", line 190, in compile
    return _compile(pattern, flags)
  File "/usr/lib64/python2.6/re.py", line 245, in _compile
    raise error, v # invalid expression
sre_constants.error: nothing to repeat

The ? after group (\s*) is what triggers the error.  Since the
group successfully matches both the emptry string and one or more
space characters, without the ?, the ? is unnecessary.  This
patch removes it.