Add command line option parsing.  Deduce cluster name.
Use argparse to parse command line options.  This allows
the user to specify the cluster name, and/or the admin_ecoch value.

If the cluster name is not supplied, then the script attempts
to deduce the cluster name from the name of one random node in
the ldev.conf file.