Add fence_powermance stonith agent support.
This change will create a stonith resource using the LLNL-developed
fence_pacemaker agent.

At this time, fence_pacemaker does not implement the "list" action,
so the configuration _must_ explicitly list all nodes in the
pcmk_host_list instance attribute of the resource.

A single stonith resource is added, and then (since we are using
pacemaker asymmetric cluster mode) we need to a resource location
constraints to tell pacemaker where it is allowed to run.  We list
all nodes as equally good places for the fence_pm resource to run.