ldev2pcs working for single managment node
This update gets ldev2pcs working on the jet testbed at LLNL.

The script assumes only a single managment node.  Powerman
works locally on that node.

MGS is given an Optional order constraint to precede all of the
other Lustre services.

Lustre services resources are marked "resource-discovery=exclusive",
which should help with start up performance to some extent.  That
makes it so that Pacemaker will only do sanity checks for the
resource's existence on the nodes where it is expected to live.
By default it would have checked on _all_ known nodes, and seemed
to check slowly and sequentially when it did.