Provide hooks pre/post start and stop
Allow the admin to specify actions to be taken before or after
the lustre agent starts its resource, or stops its resource.

The agent looks for an executable (e.g. a script) in the directory
where the agent itself is located, named
    lustre_pre_start
    lustre_post_start
    lustre_pre_stop
    lustre_post_stop

and attempts to execute the executable.  If that execution fails, or if
the exit code is nonzero, the agent fails the action.