Remove --local argument to pcs setup
The --local argument causes pcs setup to remove /etc/pacemaker/authkey.
On non-RHEL distributions, with certain arguments, 'pcs setup' will
create a new authkey and distribute it to the other pacemaker nodes.
It does not apply in our case, and removing the authkey interferes
with our cluster configuration via cfengine.