Alter lflush to issue clear to servers in parallel
The shell command that triggers lustre to flush dirty pages and drop
ldlm locks,

        echo "clear" > {namespace-server-path-to-lru_size}

blocks until the process is complete.  Since separate servers are
designed to operate independently, the servers can be done in parallel.

To tide us over while we write a similar command in lustre itself, issue
these shell commands in parallel for each of the servers, and wait for
them to finish.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>