Check /proc/spl/kstat/zfs/<POOL>/state for pool health (#12)
ZFS 0.7.10 includes a /proc file for the pool state:

$ cat /proc/spl/kstat/zfs/<POOL>/state
ONLINE

If the file exists, use it first to determine if the pool is
healthy or not.  The /proc/.../state file is lockless and will
never hang (unlike 'zpool status').  If the file isn't there
then fall back to looking for just the pool /proc file
(/proc/spl/kstat/zfs/<POOL>).

Also, set migration-threshold=1 on all zpool resources so
suspended pools get migrated immediately.

Signed-off-by: Tony Hutter <hutter2@llnl.gov>
Reviewed-by: Christopher J. Morrone <morrone2@llnl.gov>
Reviewed-by: Olaf Faaland <faaland1@llnl.gov>