Add zpool agent parameter name="post_import_cmd"
Add parameter name="post_import_cmd" to the zpool resource agent for
pacemaker.  The parameter is optional.

During resource start, after the pool is successfully imported, the
given command (if one was specified) is executed, with the pool name
passed as a command line argument.  If the command fails, determined by
nonzero exit code, log an error message, export the pool, and fail the
resource start with $OCF_ERR_GENERIC.

This allows the user to perform additional work after the resource is
started, such as setting tunables which are not persistent, or
checking pool properties to enforce a standard configuration.

If the post_import_cmd is not set in the resource configuration, the
resource agent returns OCF_SUCCESS on a successful import as in the
past.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>