Add lustre agent parameter name="post_mount_cmd"
Add parameter name="post_mount_cmd" to the lustre resource agent for
pacemaker.  The parameter is optional.

During resource start, after the target is successfully mounted, the
given command (if one was specified) is executed, with the dataset
name and mountpoint passed as command line arguments.  If the command
fails, determined by nonzero exit code, log an error message, umount
the target, and fail the resource start with $OCF_ERR_GENERIC.

This allows the user to perform additional work after the resource is
started, such as setting tunables which are not persistent, or
checking target properties to enforce a standard configuration.

If the post_mount_cmd is not set in the resource configuration, the
resource agent returns OCF_SUCCESS on a successful mount as in the
past.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>