Added some minor logging, updated log level for failure cases
Modified lustre and zpool resource agent scripts.

The purpose of this PR is to identify where in the process pacemaker is
detecting a failure and fencing nodes.  By adding logging we hope to
identify what the relevant {lustre,zpool}_monitor function thinks the
state is before reporting a failure back to Pacemaker.

I've also adjusted the log level to info so that we catch these failure
cases without being swamped with debug messages.  Specifically, this is
to help us solve an issue where Pacemaker is fencing nodes that we don't
think it should be fencing.  In most cases this should generate a small
number of additional log messages that should not affect production
environments.

addressed review comments, updated working order

Signed-off-by: Thomas Grant Bennett <bennett66@llnl.gov>
Reviewed-by: Olaf Faaland <faaland1@llnl.gov>