Ported llogcolor to python3 and added tests
llogcolor has been ported from python2 to python3.

The optparse library (which is deprecated)
has been replaced with the argparse library.

The function next_ansi_color() was removed because it
modifies global state and this is a problem for tests
that expect outputs to match verbatim. The functionality
of next_ansi_color() is now in main().

Many print statements were changed to use string formatting.

The pathlib library is being used in place of the os library
for file and directory operations.

The parser is built in a separate function.

main() is modified so as to allow arguments not just from the
command line. This is for testing purposes.

The various outputs (terminal, less, files) are all set
up to accept strings.

There is now a test directory, which contains a test for
llogcolor. It also has reference files for the test outputs.
The test can be run directly with python:
>$ python test_llogcolor.py