Additional formatting before python3 code changes
Formatted with black to conform to local style
standards.

The black formatting tool was allowed to make
all quotation characters consistent, and used
double quotes by default. Previously, there
was a mix of single and double quotes.

Alphabetized the imports. Not required but generally
good practice, and future commits will be changing
the imports so I don't want to alphabetize later.