Converted regex string to raw string
A string used for a regex was represented
as the concatenation of several strings on
different lines, however, only the first was
prefixed with an "r". Consequently only the fist
part of the string was a raw string literal. This
commit makes all of the string literals for the
regex raw strings.