Converted from python2 to python3
This commit changes the required python
interpreter from python2 to python3.

The deprecated optparse library is replaced with
the similar argparse library.

The python2 print keyword is replaced with the
print function.

A change was made so that all possible outputs
(less, the console, and files) are all opened so
that they accept strings (and not bytes). This is
due to python2 and python3 having differences in how
strings and bytes are handled.