Added testing for llogcolor
llogcolor is modified so that arguments besides those
from the command line can be used.
This feature is used by its new testing script
test_llogcolor.py.

A test directory is created with a single test script
and sample log input and output files that are used as
references for the tests.

The function next_ansi_color() was removed because it
modifies global state and this is a problem for tests
that expect outputs to match verbatim and run main() more than
once with reloading the module. The functionality of
next_ansi_color() is now in main().