Style and library changes
These changes don't change how the code works,
they are based on my personal preferences and are for
the purpose of making the code more readable.

The parser setup is removed from main() and put into
its own function. This is because main() is very long
and the parser code is self-contained.

The pathlib library is used for file and directory
operations. This replaces some uses of the os library.
This is because I find the pathlib code to be cleaner
than the corresponding os code in most cases.