Fix bug due to change in lfs output
A change in the output of "lfs quota -q <mount>" caused a bug.
lustre_quota_summary incorrectly interpreted a line
in the lfs output to be numerical quota data, which caused
the wrong data to be reported by lustre_quota_summary.
Instead of reporting the amount of bytes used, the user's
uid was incorrectly interpreted as kilobytes used.
The number of inodes used was replaced by the word "file".

In previous versions of lfs there were no lines
after the line with the quota data. The fix is to simply
stop reading input once the correct input has been found.