update and format zfsobj2fid for python3
Changed zfsobj2fid to use python3 and not python2.
Formatting tool black changed single quotes to
double quotes and put spaces around operators.

The call to zdb is changed to use subprocess.run()
and then converts the captured text to a string,
which was not necessary in python2. Also, stderr
is not captured because it was not used, but the
subprocess is now checked for a bad return value.

Long lists of arguments are put on separate lines.
Backslashes at the end of lines are made unnecessary
by wrapping the entire expression in parentheses.