update ldev2pcs for pacemaker 2.X
The command to make a corosync.conf file, as well as
a resource creation command, have a different syntax for
pacemaker 2.X vs pacemaker 1.X.
Check which version of pacemaker is being used and change the
syntax accordingly.
Additionally, there is an command line option to skip making
a new corosync.conf file.