ldev2pcs: allow time for crashing node to write out vmcore
When lustre server nodes crash due to LBUGs, BUGs, etc. pacemaker detects the
lost connection to pacemaker_remote and requests the crashing node be fenced.

When pacemaker doesn't interfere, it takes at least 45 seconds for a server on
gopher to boot into the crash kernel and write out the dump.  However, in one
test on gopher, the interval between pacemaker detecting the node was down and
fencing it was only 28 seconds.

This is consistent with production, where we almost always get incomplete
vmcore files.

Configure pacemaker to add a delay before fencing via these settings:

	pcmk_delay_base=60s
	pcmk_delay_max=90s

In a test on gopher, this allowed the dump to be written out with approximately
67 seconds to spare.  Production file systems will have much more in memory, so
dumps will take longer, but this seems more than adequate.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>