use file path for BuildRequires
instead of using lustre and lustre-devel package names
for BuildRequires, use the path to the header file
that those packages are providing. This is to get around the
issue that the header file used to be in lustre, but is now
is lustre-devel.

Also restore the header file check in configure.ac to make
rpm and non-rpm builds consistent.